requirements-production.txt
# Production Requirements for Azure Functions
# Core framework
fastapi==0.115.13
uvicorn==0.34.3
azure-functions==1.23.0

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
GeoAlchemy2==0.17.1

# Pydantic for data validation
pydantic==2.11.7
pydantic-settings==2.9.1

# HTTP clients
httpx==0.28.1
aiohttp==3.12.13

# Authentication & Security
PyJWT==2.10.1
python-multipart==0.0.20

# External services
azure-storage-blob==12.25.1
google-genai==1.20.0
supabase==2.15.3

# Configuration
python-dotenv==1.1.0

# Async support
anyio==4.9.0

# Dependencies for external packages
cryptography==45.0.4
requests==2.32.4
numpy==2.2.6 